% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\name{brute_force_knapsack}
\alias{brute_force_knapsack}
\title{Brute force knapsack
Solve the knapsack problem using brute-force search, iterating over all possible combinations.}
\usage{
brute_force_knapsack(x, W, parallel = FALSE)
}
\arguments{
\item{x}{data.frame with two variables v and w describing the values and weights of the objects to fit into the knapsack}

\item{W}{a numeric scalar that is is a positive, whole number describing knapsack size (maximum weight it can fit)}

\item{parallel}{a boolean specifying whether the search should be parallelized, defaults to FALSE}
}
\value{
a list containing value, the combined value of the elements fitted into the knapsack, and elements, the indexes of the elements that are the best fit
}
\description{
Brute force knapsack
Solve the knapsack problem using brute-force search, iterating over all possible combinations.
}
\examples{
x <- data.frame(v=c(5,4,3,2), w=c(4,3,2,1))
brute_force_knapsack(x=x, W=6)
}
