% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\name{greedy_knapsack}
\alias{greedy_knapsack}
\title{Greedy knapsack solution approximation
Use a greedy heuristic to approximate the solution to the knapsack problem.
This sorts objects by their value per weight and greedily picks the top objects until the sack is full.}
\usage{
greedy_knapsack(x, W)
}
\arguments{
\item{x}{data.frame with two variables v and w describing the values and weights of the objects to fit into the knapsack}

\item{W}{a numeric scalar that is is a positive, whole number describing knapsack size (maximum weight it can fit)}
}
\value{
a list containing value, the combined value of the elements fitted into the knapsack, and elements, the indexes of the elements that are the best fit
}
\description{
Greedy knapsack solution approximation
Use a greedy heuristic to approximate the solution to the knapsack problem.
This sorts objects by their value per weight and greedily picks the top objects until the sack is full.
}
\examples{
x <- data.frame(v=c(5,4,3,2), w=c(4,3,2,1))
greedy_knapsack(x=x, W=6)
}
