% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knapsack.R
\name{knapsack_dynamic}
\alias{knapsack_dynamic}
\title{Dynamic programming knapsack solver
Solve the knapsack problem using the dynamic programming in advance algorithm.
This iterates over all possibles values of w to find the maximum value obtainable for the weight.}
\usage{
knapsack_dynamic(x, W)
}
\arguments{
\item{x}{data.frame with two variables v and w describing the values and weights of the objects to fit into the knapsack}

\item{W}{a numeric scalar that is is a positive, whole number describing knapsack size (maximum weight it can fit)}
}
\value{
a list containing value, the combined value of the elements fitted into the knapsack, and elements, the indexes of the elements that are the best fit
}
\description{
Dynamic programming knapsack solver
Solve the knapsack problem using the dynamic programming in advance algorithm.
This iterates over all possibles values of w to find the maximum value obtainable for the weight.
}
\examples{
x <- data.frame(v=c(5,4,3,2), w=c(4,3,2,1))
knapsack_dynamic(x=x, W=6)
}
